requirements.txt
Django==5.2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
mongoengine==0.27.0
pymongo==4.6.1
django-cors-headers==4.3.1
python-dotenv==1.0.1 